split default twiss pattern in a submodule
