+++ tests/test_tfstools.py
@@ -1,7 +1,6 @@
 import pathlib
 
 import matplotlib
-import matplotlib.pyplot as plt
 import numpy as np
 import pytest
 import tfs
